+++ spec/integration/spec_helper.rb
@@ -93,7 +93,11 @@ vagrant_machines.each do |server|
                  true
                else
                  false
-               end
+               end,
+      ssh: case test_environment
+           when "staging"
+             { host_name: inventory.host(server)["ansible_host"], user: "ec2-user" }
+           end
     )
   end
 end