+++ roles.galaxy/trombik.opensmtpd/CHANGELOG.md
@@ -1,3 +1,21 @@
+## Release 1.4.0
+
+* e577eb1 [feature] Support x509-certificate (#32)
+
+## Release 1.3.2
+
+* cbca001 [bugfix] remove include: (#30)
+
+## Release 1.3.1
+
+* e2ba83e [bugfix] QA (#28)
+
+## Release 1.3.0
+
+* c858284 [bugfix] Notify the daemon when file databases have been updated (#26)
+* 7d3b558 [feature] Support table types other than file and db (#25)
+* 81b281d [bugfix] support Ubuntu (#22)
+
 ## Release 1.2.0
 
 * 4045034 [feature] support `opensmtpd_extra_groups` (#20)